[expo-cli] Validate the credentials for android keystore
Why
===
We want to ensure that the user provided correct keystore file and credentials before starting the build. It'll let us avoid cases such as when users downloade some file from Google play and mistakenly thought that it's a keystore file.

How
===
1. When reading the credentials for Android from 'credentials.json', validate the keystore file with 'keytool' by using the provided credentials, unless '--skip-credentials-check' is specified.
2. If 'keytool' is not present, print an warning and skip the verification.

Test plan
=========
Specify incorrect credentials in 'credentials.json' and try to trigger a build, and make sure that the an error is thrown if the keystore couldn't be validated with 'keytool'.

Tested with the 'keytool' from both Oracle JDK and OpenJDK to make sure it works in both cases.